added colors and opacity and corrections in radius

+++ src/main/java/com/necleo/codemonkey/lib/types/figma/properties/fills/Fills.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 public class Fills {
   public String type;
   public boolean visible;
-  public int opacity;
+  public float opacity;
   public String BlendMode;
   public Color color;
 }
